Documents/doc_search_prototype2/requirements.txt
fastapi
uvicorn
streamlit
pdfplumber
faiss-cpu
sentence-transformers
numpy
requests
watchdog
tiktoken
python-dotenv
PyPDF2
docx2txt